sched_policy: fix set_timerslack_ns function prctl error

+++ libcutils/sched_policy.c
@@ -345,7 +345,7 @@ static void set_timerslack_ns(int tid, unsigned long long slack) {
     }
 
     /* If the above fails, try the old common.git PR_SET_TIMERSLACK_PID. */
-    if (prctl(PR_SET_TIMERSLACK_PID, slack, tid) == -1) {
+    if (prctl(PR_SET_TIMERSLACK_PID, (unsigned long)slack, tid) == -1) {
         SLOGE("set_timerslack_ns prctl failed: %s\n", strerror(errno));
     }
 }
